Add data transfer between the client, the map and the api
Before this change, the data was statically loaded from within the JavaScript code.
This has now been changed, the data is fetched directly from the API.

This has been done for quality and maintenance purposes, since most of the application is written in C#.
It also adds a layer of security, since the Mapbox token is now loaded in through application settings, instead of being hardcoded.